requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
cffi==1.11.5
Click==7.0
colorama==0.4.1
coverage==4.5.2
cryptography==2.4.2
flasgger==0.9.2
Flask==1.0.2
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
more-itertools==5.0.0
pep8==1.7.1
pluggy==0.8.1
psycopg2==2.7.6.1
py==1.7.0
pycodestyle==2.4.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.2.2
pytest==4.1.1
pytest-cov==2.6.1
PyYAML==3.13
six==1.12.0
Werkzeug==0.14.1
wrapt==1.11.1
